Correct insertion of large string input. (Paste)
It check each character of the paste input one by one to check if it is
valid.
Also if selected a range of characters and put a new input it deletes
the selected area and checks if the new input will be correct.
Delete is always allowed. (Might lead to bugs of deleting something and
having a wrong string to calculate).