Solve operations
Solve operations based on strings. This operations can contain:

- '+'

- '-'

- '/'

- '*'

- '(' and ')'

Interface to input characters to a string and validate if the insertion
would create a valid operation. Drawback of this method is, that if the
user made a mistake and forgot to put a number or operator it has to
first add the number and then put the operator.